requirements.txt
numpy
opencv-python-headless
scikit-image
tifffile
Pillow
scipy
matplotlib
scikit-learn
joblib
